requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0
facenet-pytorch>=2.5.3
mtcnn>=0.1.1

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Vector Search
faiss-cpu>=1.7.4  # Use faiss-gpu for GPU support
# faiss-gpu>=1.7.4  # Uncomment for GPU

# API Server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Experiment Tracking
wandb>=0.15.0
# mlflow>=2.5.0  # Alternative to wandb

# Configuration
PyYAML>=6.0
python-dotenv>=1.0.0
hydra-core>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
tensorboard>=2.13.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
loguru>=0.7.0
python-json-logger>=2.0.7

# Data Versioning
dvc>=3.0.0
dvc-s3>=3.0.0  # For S3 remote storage

# Image Quality Assessment
brisque>=0.0.3
piq>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.4.0

# Production
gunicorn>=21.2.0
prometheus-client>=0.17.0
sentry-sdk>=1.28.0

# Qwen3-VL and Transformers (Required for SageMaker)
transformers>=4.45.0
qwen-vl-utils>=0.0.8
accelerate>=0.24.0
sentencepiece>=0.1.99
einops>=0.7.0

# SageMaker SDK
sagemaker>=2.200.0
boto3>=1.34.0
botocore>=1.34.0

# Optional: Advanced Features
# timm>=0.9.0  # PyTorch Image Models
# onnx>=1.14.0  # Model export
# onnxruntime>=1.15.0  # ONNX inference
